fix: price Long -> Double 대대적 변경 수정

+++ backend/funco-follow/src/main/java/com/found_404/funco/feignClient/service/TradeService.java
@@ -40,7 +40,7 @@ public class TradeService {
 		}
 	}
 
-	public Map<String, Long> getCryptoPrice(List<String> tickers) {
+	public Map<String, Double> getCryptoPrice(List<String> tickers) {
 		try {
 			return tradeServiceClient.getCryptoPrice(tickers);
 		} catch (FeignException e) {
